Fix GCC warnings (#554)
* constraintsbuilder.cpp : remove unused arguments

* Remove useless asserts &variable != NULL

* Fix compilation warnings in studylogs.{h/cpp}

* Fix compilation warnings in exportmap.cpp

* Remove unused asserts & function parameters

* Fix several compilation warnings

* control.{h,cpp} : Use in-class initializer

* Pass const char* instead of char*

* Remove unused variables

* Fix more compilation warnings

* Fix more compilation warnings

* Fix CMake

* Restore STUDY_MODE_EXPANSION

* Fix more compilation warnings

* Make constants 'const'

* Add 'explicit' keyword

* Revert change in yuni job queue

* PR remarks

* PR remarks

* Fix warnings in src/tests/libs/antares

* Fix possibily unitialized variable

* Remove unused 'warn' argument from Notebook::find

* Use std::size_t in tests

* Remove warning in debug mode

It wasn't displayed beforehand.

* Remove unused variable MapPage

* Remove commented-out code

+++ src/ui/simulator/toolbox/components/datagrid/renderer/area/thermalmodulation.cpp
@@ -121,7 +121,7 @@ double ThermalClusterCommonModulation::cellNumericValue(int x, int y) const
 {
     if (pMatrix && (uint)y < pMatrix->height && pCluster)
     {
-        assert((uint)x < static_cast<int>(width()));
+        assert(x < width());
         switch (x)
         {
         case Data::thermalModulationCost:
